+++ src/settings.cpp
@@ -40,29 +40,10 @@ void wps::setparams( const wps_parameters params )
     _settings.set( params, get_self() );
 }
 
-[[eosio::action]]
-void wps::pause( const bool paused )
-{
-    require_auth( get_self() );
-
-    auto settings = _settings.get();
-
-    check( settings.paused != paused, "nothing modified");
-    settings.paused = paused;
-    _settings.set( settings, get_self() );
-}
-
-void wps::check_contract_paused()
-{
-    auto settings = _settings.get();
-    check( !settings.paused, "contract is currently paused");
-}
-
 void wps::check_contract_active()
 {
     check( _state.exists(), "contract not yet initialized" );
     check( _settings.exists(), "settings are missing" );
-    check_contract_paused();
 }
 
 void wps::add_funding( const asset quantity )